pyproject.toml
[tool.poetry]
name = "main"
version = "0.1.0"
description = ""
authors = ["diamondburned <diamond@arikawa-hi.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4"
aiosqlite = "0.19.0"
annotated-types = "0.6.0"
anyio = "4.2.0"
fastapi = "0.109.0"
h11 = "0.14.0"
httptools = "0.6.1"
idna = "3.6"
iniconfig = "2.0.0"
pluggy = "1.4.0"
pydantic = "2.6.0"
pydantic-core = "2.16.1"
pytest = "7.3.1"
pytest-asyncio = "0.21.0"
python-dotenv = "1.0.1"
pyyaml = "6.0.1"
sniffio = "1.3.0"
sqlalchemy = "2.0.25"
sqlmodel = "0.0.14"
starlette = "0.35.1"
typing-extensions = "4.9.0"
uvicorn = "0.27.0.post1"
uvloop = "0.19.0"
watchfiles = "0.21.0"
websockets = "12.0"
sonyflake-py = "^1.3.0"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
pyright = "^1.1.349"

[tool.pyright]
typeCheckingMode = "strict"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
